Re-arranged examples
