+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/CredentialConfiguration.kt
@@ -29,6 +29,13 @@ import java.util.*
 sealed interface CryptographicBindingMethod : Serializable {
 
     /**
+     * X5C format.
+     */
+    data object X5C : CryptographicBindingMethod {
+        private fun readResolve(): Any = X5C
+    }
+
+    /**
      * JWK format.
      */
     data object JWK : CryptographicBindingMethod {
